Removing STSS_wrapper.py
The core functionalities this script was used for was ported to the main STSS.py script. The only remaining functionality for STS annotating that is not in STSS is the ability to provide a list of assembly uIDs to exclude from searching. That still remains in STSS_auto_update.py, and could be ported, but it isn't that important except for maintaining the whole database from NCBI. Exclusion can also be done manually by taking the uid download list file and removing unwanted uIDs with simple script.